packages/crawler/requirements.txt
# 公考职位智能筛选系统 - 爬虫模块依赖
# Python >= 3.10

# 爬虫框架
scrapy>=2.11.0
selenium>=4.15.0
playwright>=1.40.0
requests>=2.31.0
httpx>=0.25.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0
parsel>=1.8.0

# 文档解析
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
xlrd>=2.0.0

# OCR识别
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0

# AI/LLM
openai>=1.3.0
anthropic>=0.7.0
tiktoken>=0.5.0

# 数据库
pymysql>=1.1.0
sqlalchemy>=2.0.0
redis>=5.0.0
elasticsearch>=8.11.0

# 任务队列
celery>=5.3.0
celery[redis]>=5.3.0
flower>=2.0.0

# GUI界面 (WebView)
pywebview>=4.4.0

# API服务
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
tenacity>=8.2.0
fake-useragent>=1.4.0
python-dateutil>=2.8.0

# 代理和反爬
aiohttp>=3.9.0
aiohttp-socks>=0.8.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 开发工具
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
